+++ docs-requirements.txt
@@ -48,6 +48,11 @@ starlette~=0.50
 tornado>=5.1.1
 tortoise-orm>=0.17.0
 
+# required by opamp
+uuid_utils
+protobuf>=5.0,< 7.0
+
 # indirect dependency pins
 markupsafe==2.0.1
 itsdangerous==2.0.1
+
